% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lcsubstr.R
\name{find_lcsubstr}
\alias{find_lcsubstr}
\title{Find the longest common substring between elements of a character vector}
\usage{
find_lcsubstr(string, n_min = 1, n_max = NULL, trace = TRUE)
}
\arguments{
\item{string}{a character vector.}

\item{n_min}{integer. Minimum common substring length to find.}

\item{n_max}{integer. Maximum common substring length to find. Default (\code{NULL}) correspond to the length of the smallest string}

\item{trace}{logical. Should the result be printed in the console ?}
}
\value{
a character vector with the longest common substring(s)
}
\description{
Find the longest common substring between elements of a character vector
}
\examples{
st1 <- "Here is a first string for the example, how original !"
st2 <- "Here a second string for the example too."
st3 <- "Here is a third  string for the example, as original as the others."
find_lcsubstr(string = c(st1, st2, st3))
}
