% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcsubstr-package.R
\docType{package}
\name{lcsubstr-package}
\alias{lcsubstr}
\alias{lcsubstr-package}
\title{lcsubstr: Find the longest common substring between elements of a character vector}
\description{
This package contains only one function which find the longest
common substring between elements of a character vector.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BenjaminLouis/lcsubstr/}
  \item Report bugs at \url{https://github.com/BenjaminLouis/lcsubstr/issues/}
}

}
\author{
\strong{Maintainer}: Benjamin Louis \email{benjaminlouis.pro@gmail.com}

}
\keyword{internal}
